+++ src/main/java/ui/controller/BoardController.java
@@ -20,6 +20,7 @@ package ui.controller;
 import event.EventBus;
 import event.GameEvent;
 import event.ScoreEvent;
+import event.decorators.GameHandlerEventDecorator;
 import javafx.beans.value.ObservableValue;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
@@ -35,7 +36,6 @@ import javafx.scene.layout.VBox;
 import javafx.stage.FileChooser;
 import logic.board.BoardScorer;
 import logic.gamehandler.GameHandler;
-import logic.gamehandler.GameHandlerEventDecorator;
 import logic.gamehandler.LocalGameHandler;
 import sgf.SGFWriter;
 import ui.drawer.BoardDrawer;
